+++ lib/pagy.rb
@@ -22,6 +22,7 @@ class Pagy
               fragment:      '',
               link_extra:    '',
               item_i18n_key: 'pagy.item_name',
+              page_i18n_key: 'pagy.page_label',
               cycle:         false,
               request_path:  '' }