+++ lib/pagy/extras/frontend_helpers.rb
@@ -6,7 +6,7 @@ class Pagy # :nodoc:
   # Private module documented in the main classes
   module FrontendHelpers
     # Additions for the Pagy class
-    module Pagy
+    module PagyOverride
       # `Pagy` instance method used by the `pagy*_nav_js` helpers.
       # It returns the sequels of width/series generated from the :steps hash
       # Example:
@@ -27,9 +27,10 @@ class Pagy # :nodoc:
       # Support for the Calendar API
       def label_sequels(*); end
     end
+    Pagy.prepend PagyOverride
 
     # Additions for Calendar class
-    module Calendar
+    module CalendarOverride
       def label_sequels(sequels = self.sequels)
         {}.tap do |label_sequels|
           sequels.each do |width, series|
@@ -38,9 +39,10 @@ class Pagy # :nodoc:
         end
       end
     end
+    Calendar.prepend CalendarOverride if defined?(Calendar)
 
     # Additions for the Frontend
-    module Frontend
+    module FrontendOverride
       if defined?(Oj)
         # Return a data tag with the base64 encoded JSON-serialized args generated with the faster oj gem
         # Base64 encoded JSON is smaller than HTML escaped JSON
@@ -65,8 +67,6 @@ class Pagy # :nodoc:
         link.call PAGE_PLACEHOLDER, '', 'style="display: none;"'
       end
     end
+    Frontend.prepend FrontendOverride
   end
-  prepend FrontendHelpers::Pagy
-  Calendar.prepend FrontendHelpers::Calendar if defined?(Calendar)
-  Frontend.prepend FrontendHelpers::Frontend
 end